requirements.txt
biopython==1.81
numpy==1.23.5
pandas==2.1.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
umap-learn==0.5.5
streamlit==1.29.0
joblib==1.3.2
tqdm==4.66.1
